- implement api versioning - move and define controllers to v1 namespace - move recipe resource to v1 namespace
