+++ config/environments/development.rb
@@ -1,3 +1,4 @@
+require 'api_version'
 Rails.application.configure do
   # Settings specified here will take precedence over those in config/application.rb.
